PWA-1089 Corrected styles
