+++ libraries/ui-ios/AppBar/style.js
@@ -12,6 +12,7 @@ const outer = css({
 
 const inner = css({
   display: 'flex',
+  justifyContent: 'space-between',
   position: 'relative',
   zIndex: 1,
 });
