--- libraries/ui-ios/Notification/transition.js
@@ -1,8 +1,0 @@
-export default {
-  exiting: {
-    transform: 'translate3d(0, -100%, 0)',
-  },
-  exited: {
-    transform: 'translate3d(0, -100%, 0)',
-  },
-};
